# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: The Nexus Engine - Prebid Server API
  description: |
    OpenRTB 2.5 compliant Prebid Server for programmatic advertising auctions.

    ## Authentication
    API requests require authentication via API key, passed in either:
    - `X-API-Key` header
    - `Authorization: Bearer <api_key>` header

    Public endpoints (`/health`, `/status`, `/metrics`, `/info/bidders`) do not require authentication.

    ## Rate Limiting
    Requests are rate-limited per publisher. When rate-limited, the API returns 429 Too Many Requests.
  version: 1.0.0
  contact:
    name: The Nexus Engine Team
    url: https://github.com/StreetsDigital/thenexusengine
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT

servers:
  - url: https://api.thenexusengine.com
    description: Production server
  - url: http://localhost:8000
    description: Local development

tags:
  - name: Auction
    description: OpenRTB auction endpoints
  - name: Info
    description: Server information endpoints
  - name: Health
    description: Health check endpoints
  - name: Admin
    description: Administrative endpoints

paths:
  /openrtb2/auction:
    post:
      tags:
        - Auction
      summary: Run OpenRTB 2.5 Auction
      description: |
        Executes a real-time bidding auction according to OpenRTB 2.5 specification.
        The request is sent to configured bidders, bids are collected and validated,
        and the winning bid is returned.
      operationId: runAuction
      security:
        - ApiKeyAuth: []
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              \$ref: '#/components/schemas/BidRequest'
            examples:
              banner:
                summary: Banner ad request
                value:
                  id: "auction-123"
                  imp:
                    - id: "imp-1"
                      banner:
                        w: 300
                        h: 250
                      bidfloor: 0.50
                  site:
                    page: "https://example.com/article"
                    publisher:
                      id: "pub-123"
              video:
                summary: Video ad request
                value:
                  id: "auction-456"
                  imp:
                    - id: "imp-1"
                      video:
                        mimes: ["video/mp4"]
                        w: 640
                        h: 480
                        minduration: 5
                        maxduration: 30
                      bidfloor: 2.00
                  site:
                    page: "https://example.com/video"
      responses:
        '200':
          description: Successful auction response
          headers:
            X-Request-ID:
              schema:
                type: string
              description: Unique request identifier for tracing
          content:
            application/json:
              schema:
                \$ref: '#/components/schemas/BidResponse'
        '204':
          description: No bid - auction completed with no winning bids
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                \$ref: '#/components/schemas/Error'
        '401':
          description: Missing API key
          content:
            application/json:
              schema:
                \$ref: '#/components/schemas/Error'
        '403':
          description: Invalid API key
          content:
            application/json:
              schema:
                \$ref: '#/components/schemas/Error'
        '413':
          description: Request too large
          content:
            application/json:
              schema:
                \$ref: '#/components/schemas/Error'
        '429':
          description: Rate limit exceeded
          content:
            application/json:
              schema:
                \$ref: '#/components/schemas/Error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                \$ref: '#/components/schemas/Error'

  /health:
    get:
      tags:
        - Health
      summary: Health check
      description: Returns server health status. No authentication required.
      operationId: healthCheck
      responses:
        '200':
          description: Server is healthy
          content:
            application/json:
              schema:
                \$ref: '#/components/schemas/HealthResponse'

  /status:
    get:
      tags:
        - Health
      summary: Status check
      description: Returns basic server status. No authentication required.
      operationId: statusCheck
      responses:
        '200':
          description: Server is running
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string

  /info/bidders:
    get:
      tags:
        - Info
      summary: List available bidders
      description: Returns a list of all available bidders. No authentication required.
      operationId: listBidders
      responses:
        '200':
          description: List of available bidders
          content:
            application/json:
              schema:
                type: object
                properties:
                  bidders:
                    type: array
                    items:
                      type: string

  /metrics:
    get:
      tags:
        - Health
      summary: Prometheus metrics
      description: Returns server metrics in Prometheus format. No authentication required.
      operationId: getMetrics
      responses:
        '200':
          description: Prometheus metrics
          content:
            text/plain:
              schema:
                type: string

  /admin/circuit-breaker:
    get:
      tags:
        - Admin
      summary: Circuit breaker status
      description: Returns the current state of the IDR service circuit breaker.
      operationId: getCircuitBreakerStatus
      security:
        - ApiKeyAuth: []
        - BearerAuth: []
      responses:
        '200':
          description: Circuit breaker status
          content:
            application/json:
              schema:
                type: object
                properties:
                  state:
                    type: string
                    enum: [closed, open, half-open]
                  failures:
                    type: integer
                  last_failure:
                    type: string
                    format: date-time

components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: API key for authentication
    BearerAuth:
      type: http
      scheme: bearer
      description: Bearer token authentication

  schemas:
    BidRequest:
      type: object
      required:
        - id
        - imp
      properties:
        id:
          type: string
          description: Unique ID of the bid request
        imp:
          type: array
          minItems: 1
          items:
            \$ref: '#/components/schemas/Impression'
          description: Array of impression objects
        site:
          \$ref: '#/components/schemas/Site'
        app:
          \$ref: '#/components/schemas/App'
        device:
          \$ref: '#/components/schemas/Device'
        user:
          \$ref: '#/components/schemas/User'
        test:
          type: integer
          enum: [0, 1]
          description: Indicator of test mode
        at:
          type: integer
          enum: [1, 2]
          default: 2
          description: Auction type (1=first price, 2=second price)
        tmax:
          type: integer
          description: Maximum time in milliseconds to wait for bids
        cur:
          type: array
          items:
            type: string
          description: Allowed currencies
        regs:
          \$ref: '#/components/schemas/Regs'
        ext:
          type: object

    Impression:
      type: object
      required:
        - id
      properties:
        id:
          type: string
        banner:
          \$ref: '#/components/schemas/Banner'
        video:
          \$ref: '#/components/schemas/Video'
        native:
          type: object
        audio:
          type: object
        bidfloor:
          type: number
          format: float
        bidfloorcur:
          type: string
          default: "USD"
        secure:
          type: integer
          enum: [0, 1]
        ext:
          type: object

    Banner:
      type: object
      properties:
        w:
          type: integer
        h:
          type: integer
        format:
          type: array
          items:
            type: object
            properties:
              w:
                type: integer
              h:
                type: integer
        pos:
          type: integer

    Video:
      type: object
      properties:
        mimes:
          type: array
          items:
            type: string
        minduration:
          type: integer
        maxduration:
          type: integer
        w:
          type: integer
        h:
          type: integer
        protocols:
          type: array
          items:
            type: integer

    Site:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        domain:
          type: string
        page:
          type: string
        publisher:
          \$ref: '#/components/schemas/Publisher'

    App:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        bundle:
          type: string
        publisher:
          \$ref: '#/components/schemas/Publisher'

    Publisher:
      type: object
      properties:
        id:
          type: string

    Device:
      type: object
      properties:
        ua:
          type: string
        ip:
          type: string
        geo:
          \$ref: '#/components/schemas/Geo'
        devicetype:
          type: integer
        os:
          type: string

    Geo:
      type: object
      properties:
        country:
          type: string
        region:
          type: string
        city:
          type: string

    User:
      type: object
      properties:
        id:
          type: string
        buyeruid:
          type: string
        consent:
          type: string

    Regs:
      type: object
      properties:
        coppa:
          type: integer
          enum: [0, 1]
        gdpr:
          type: integer
          enum: [0, 1]
        us_privacy:
          type: string
        gpp:
          type: string
        gpp_sid:
          type: array
          items:
            type: integer

    BidResponse:
      type: object
      required:
        - id
      properties:
        id:
          type: string
        seatbid:
          type: array
          items:
            \$ref: '#/components/schemas/SeatBid'
        cur:
          type: string
        ext:
          type: object

    SeatBid:
      type: object
      properties:
        bid:
          type: array
          items:
            \$ref: '#/components/schemas/Bid'
        seat:
          type: string

    Bid:
      type: object
      required:
        - id
        - impid
        - price
      properties:
        id:
          type: string
        impid:
          type: string
        price:
          type: number
          format: float
        adm:
          type: string
        adomain:
          type: array
          items:
            type: string
        crid:
          type: string
        w:
          type: integer
        h:
          type: integer

    HealthResponse:
      type: object
      properties:
        status:
          type: string
          enum: [healthy, unhealthy]
        timestamp:
          type: string
          format: date-time
        version:
          type: string

    Error:
      type: object
      properties:
        error:
          type: string